requirements.txt
astroid==2.1.0
autopep8==1.4.3
certifi==2018.11.29
chardet==3.0.4
idna==2.8
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pycodestyle==2.5.0
pylint==2.2.2
python-forecastio==1.4.0
requests==2.21.0
responses==0.10.5
schedule==0.6.0
six==1.12.0
slackclient==1.0.0
urllib3==1.24.1
websocket-client==0.54.0
wrapt==1.11.1
